P2PNodeDeployer is no longer available to non-test code.
