+++ flocker/node/_deploy.py
@@ -903,50 +903,3 @@ def find_dataset_changes(hostname, current_state, desired_state):
                    if dataset.deleted)
     return DatasetChanges(going=going, coming=coming, deleting=deleting,
                           creating=creating, resizing=resizing)
-
-
-class P2PNodeDeployer(_OldToNewDeployer):
-    """
-    Combination of ZFS and container deployer.
-
-    Temporary expedient for use by flocker-changestate until we rip it
-    out, and even more temporarily in flocker-zfs-agent. In FLOC-1554
-    flocker-zfs-agent will be split up and stop using this.
-    """
-    def __init__(self, hostname, volume_service, docker_client=None,
-                 network=None):
-        self.manifestations_deployer = P2PManifestationDeployer(
-            hostname, volume_service)
-        self.applications_deployer = ApplicationNodeDeployer(
-            hostname, docker_client, network)
-        self.hostname = hostname
-        self.volume_service = self.manifestations_deployer.volume_service
-        self.docker_client = self.applications_deployer.docker_client
-        self.network = self.applications_deployer.network
-
-    def discover_local_state(self, local_state):
-        d = self.manifestations_deployer.discover_state(local_state)
-
-        def got_manifestations_state(manifestations_state):
-            manifestations_state = manifestations_state[0]
-            app_discovery = self.applications_deployer.discover_state(
-                manifestations_state)
-            app_discovery.addCallback(
-                lambda app_state: app_state[0].set(
-                    "manifestations", manifestations_state.manifestations).set(
-                    "paths", manifestations_state.paths))
-            return app_discovery
-        d.addCallback(got_manifestations_state)
-        return d
-
-    def calculate_necessary_state_changes(
-            self, local_state, configuration, cluster_state):
-        """
-        Combine changes from the application and ZFS agents.
-        """
-        return Sequentially(changes=[
-            self.applications_deployer.calculate_changes(
-                configuration, cluster_state),
-            self.manifestations_deployer.calculate_changes(
-                configuration, cluster_state),
-        ])
